requirements.txt
torch==2.5.1
torchaudio==2.5.1
numpy==2.0.2
accelerate==1.1.1
safetensors==0.4.5
librosa==0.10.2.post1
tqdm==4.67.1
tensorboard==2.18.0
pytest==8.3.3
einops==0.8.0
matplotlib>=3.7.0
wandb>=0.15.0
python-dotenv==1.0.0